Create HA button-card code
When the valve is open, its icon mdi:valve-open should indicate opened position, with gold color. When it is closed, the same icon should be slowly (10 sec) rotated 90 deg right, with dark grey color, and remain steady with its new rotated position at the end of the animation. Please, kindly advise how to get the following visual effect:
1. Toggling the valve on the HA dashboard (from CLOSE towards OPEN position) should trigger a rotation of the icon from CLOSE (dark grey 90 deg) to OPEN (gold 0 deg) position. 
2. Toggling the valve on the HA dashboard (from OPEN towards CLOSE position) should trigger a reversed rotation of the icon from OPEN (gold 0 deg) to CLOSE (dark grey 90 deg) position. 
Would you be so kind as to indicate what is wrong with my code, please? It still doesn't work as expected. I'm stuck..